refact(export): Move tasks export into a messageHandler
